Updated to avoid print format warnings in extended precision

+++ examples/nvector/sycl/test_nvector_sycl.cpp
@@ -441,7 +441,7 @@ int check_ans(sunrealtype ans, N_Vector X, sunindextype length)
     if (failure += SUNRCompare(Xdata[i], ans))
     {
       printf("check_ans fail: Xdata[%ld] = %f, expected Xdata[%ld] = %f\n",
-             (long int)i, Xdata[i], (long int)i, ans);
+             (long int)i, (double) Xdata[i], (long int)i, (double) ans);
     }
   }
